+++ CHANGELOG.md
@@ -9,6 +9,9 @@ glance - what to expect from upgrading to a new version.
 ### Added
 
 - Background job to fix players in state where no future updates might happen.
+- Use player's "last active at" timestamp as secondary order parameter in
+  search results. Aliases which are equally similar to the search parameter
+  will hence be sorted by the player's last activity.
 
 ### Changed